magit-status: easier customization
Make it easier to know the effect of the customization by having both
the name of the function and its documentation string displayed in the
customization screen.

Also, the section inserters defined by magit are included in a drop-down
list to ease choosing them.

The name of the customization variable is changed to better reflect the
new customization type and the actual role it has.